Refactor project to workspaces and fix limit order worker
- Created `shared` workspace containing database schema, migrations, types, and utils.
- Moved `drizzle-orm` and `drizzle-kit` logic to `shared`.
- Updated `bot/package.json` to depend on `shared` workspace.
- Updated `docker-compose.yaml` to build from root context and run migrations from `shared` workspace.
- Updated `bot/Dockerfile` to support workspace build.
- Refactored `limitOrderWorker.ts` into `OrderWorker` (orchestrator) and `PriceMonitor` (logic).
- Implemented safety check for `priceMonitor.checkPendingLimitOrders` in `OrderWorker`.
- Fixed imports in `bot` services and tests to use `shared` module.
- Ensured all tests pass.
