+++ bot/package.json
@@ -1,29 +1,19 @@
 {
   "name": "bot",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
   "scripts": {
     "build": "tsc",
     "start": "node dist/bot.js",
     "dev": "ts-node-dev -r dotenv/config --respawn --transpile-only src/bot.ts",
     "test": "jest",
-    "db:gen": "drizzle-kit generate",
-    "db:migrate": "drizzle-kit migrate",
-    "db:studio": "drizzle-kit studio",
-    "db:push": "drizzle-kit push",
     "smoke": "node smoke-test.js"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "type": "commonjs",
   "dependencies": {
-    "@neondatabase/serverless": "^1.0.2",
+    "shared": "*",
     "axios": "^1.12.2",
     "better-sqlite3": "^12.4.1",
     "dotenv": "^17.2.3",
-    "drizzle-orm": "^0.44.7",
     "ethers": "^6.13.1",
     "express": "^5.1.0",
     "groq-sdk": "^0.33.0",
@@ -35,7 +25,6 @@
     "@types/express": "^5.0.6",
     "@types/jest": "^29.5.14",
     "@types/node": "^24.10.11",
-    "drizzle-kit": "^0.31.7",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.4",
     "ts-node-dev": "^2.0.0",
