+++ shared/drizzle.config.d.ts
@@ -1,0 +1,2 @@
+declare const _default: import("drizzle-kit").Config;
+export default _default;
