+++ shared/drizzle/drizzle/0006_light_ozymandias.sql
