+++ shared/src/utils/safeParse.ts
